Implement a Sequence based on characteristic words
this implementation is a rough cut. But it works along the following lines:

1. if there is a minimal expresssion then

a. if the minimal is bigger then the candidate, candidate is unexpressable.
b. if the minimal equals the candidate, candidate is expressable
c. if the minimal is less then the candidate, we are unsure if the candidate is expressable

2. if there isn't, candidate is unexpressable

If a candidate is unexpressible, report is as the next iterator element, add it to the weights, and add new expressions
If a candidate is expressable, we can progress the minimal expression, and skip the candidate

If it is unknown if the candidate is expressable, progress the minimal expression.